Initial Commit
Learning about Bedrock's Script API, and I was curious about what kinds of APIs the runtime supports, and what modules are available in the client. This seems to be the subset of what is available. I want to ensure that it's possible to have a type-safe API surface before trying to write scripts there. This is very cool!

https://www.npmjs.com/search?q=%40minecraft (I just installed all of the available modules (that worked), so then we can get full API coverage for what seems to be available in any addon that wants to use them.)