QuickJS Typings Removal
Only certain features from the QuickJS standard library are available, and the ones defined in the typings from this package weren't anything that appears to be available in this version of the runtime. Going to implement my own typings file with the ones that do happen to be supported, like the four `console` methods, which are usually part of the DOM typings (also not available here, understandably).

Still have to deduce whether support is indeed up to upstream QuickJS' APIs with ES2023, or if it is only up to ES2020.

I'm curious how it handles module resolution under the hood too, that could be a good experiment to look into too.