Dependencies Debug (We're working now!)
Had to update the dependencies versions to ones that are specific to what the game numbers them by (they are different numbers than the npm versions), and I removed the non-client modules, as looks like the other ones are regular JS, rather than ones you import from the game itself. So this is the most functioning one I have so far! Heck yeah! Now I can start debugging the typings for what the environment supports.

`/reload` is awesome by the way.