Environment Debug
Scoping out the rest of the standard library globals. Looks like in 1.21.31 BigInt isn't supported, but in 1.20.72, it's not. So I wonder if they raised the target up from ES2020.

https://wiki.bedrock.dev/scripting/api-environment.html#support
https://help.minecraft.net/hc/en-us/articles/30298767427597 (When trying the new update, looks like the Unix launcher might have to get a big rework, or it will stop working for modern versions, as they are dropping support for old Android devices, and that's what the VM runs on)

So now I think we have a better idea what these types should look like! I think I will have to build a custom definitions file with all of the lib types, and to exclude the ones with the unsupported features mentioned here.