Full Environment Typings!
The only thing here is that BigInt shouldn't be accessible, at least for this build. I might leave it just how it is here though, since the more recent versions do support it. Not sure. Probably will still try to see if I can hand-pick ES2020 features to not have BigInt in it. Probably not too hard to get set up.