BigInt + Intl Removal (Almost)
This is how I figured out how to remove the unsupported features from the standard library.

This doesn't completely seem to be possible though, as some features are defined in the same files, and you have to hand-pick lib versions, for ones that the runtime does support. So Intl goes back to ES5, meaning you have to pick all of the other linked definitions back to that time, just omitting Intl (same with BigInt). But, turns out, these are part of the initial typings themselves, so you can't really remove them unfortunately. So, afterall, I might just use ES2020 as the typings version, and mark these APIs as 'deprecated' just so they have an additional note about being unsupported in this context. I don't want to write unmaintainable lib typings since these are changing over time anyways.

So yeah, BigInt support comes in the newer game engines, Intl isn't, and Atomics isn't. The typings need SharedArrayBuffer though, and that's in the same file as Atomics. So there isn't an in-place way to remove one or import the other, that I'm aware of. I wish you could do that! Maybe I can figure some other kind of definitions global scope modification hacks or something (at the type-level of course).