Hand-Picked Types Revert
Using deprecated tags on pre-defined APIs doesn't appear to work, which is also a bummer. So these will be a few inconsistencies in the typings of the environment. It's not that many though, so it's fairly strict, at least much more than compared to the original TS configurations and typings :)