Module Resolution Configs
Had to fix the usage of the module configuration, as it shouldn't support top-level await, but it should be able to import from node_modules. It's hard to configure TS to do that, I wish it was easier to handle things like that explicitly. Same with file extensions. I just want ES2020, with explicit file extensions, without top-level await. So essentially, NodeNext, but without top-level await. Why isn't it possible to configure things like that?