Demo Concept
This is a random thing, where it just logs out the Dimension that each player is in! Just trying to wrap my head around how this could be used for things, and how TS/JS tooling can make this very productive. Same with libraries too, how different existing modules might allow this to be a very cool thing, even cooler than it already is!